requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.12.0
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
flake8==3.9.2
flake8-docstrings==1.6.0
isort==5.12.0
mccabe==0.6.1
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4
telegram==0.0.1
tornado==6.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
